refactor(nvisy-nlp): rename NlpEngine to Engine, split language types, fold engine module
Engine module is now a single file (engine/mod.rs holds Engine; builder
stays in engine/builder.rs). LanguageDetection (with LanguageProvenance)
and LanguageSpan moved out of language/mod.rs into their own files so the
module root only carries the LanguageDetector trait + re-exports.

Engine name collides with nvisy_engine::Engine at use sites; callers
importing both will need to alias one.
