+++ crates/nvisy-nlp/src/engine/mod.rs
@@ -1,16 +1,120 @@
-//! [`NlpEngine`] — composes a [`NerBackend`] and a [`LanguageDetector`]
+//! [`Engine`] — composes a [`NerBackend`] and a [`LanguageDetector`]
 //! (both required) with an optional [`Tokenizer`] into a single
-//! entrypoint, plus the [`NlpEngineBuilder`] that constructs it.
+//! entrypoint, plus the [`EngineBuilder`] that constructs it.
 //!
 //! [`NerBackend`]: crate::ner::NerBackend
 //! [`LanguageDetector`]: crate::language::LanguageDetector
 //! [`Tokenizer`]: crate::tokenizer::Tokenizer
 
 mod builder;
-mod nlp_engine;
 
-pub use self::builder::{NlpEngineBuilder, NoLang, NoNer, WithLang, WithNer};
-pub use self::nlp_engine::NlpEngine;
+pub use self::builder::{EngineBuilder, NoLang, NoNer, WithLang, WithNer};
+
+use std::collections::HashSet;
+use std::sync::Arc;
+
+use nvisy_ontology::primitive::LanguageTag;
+
+use crate::artifacts::{Artifacts, Token};
+use crate::error::Result;
+use crate::language::{LanguageDetection, LanguageDetector, LanguageProvenance};
+use crate::ner::NerBackend;
+use crate::tokenizer::Tokenizer;
+
+/// Composite NLP engine.
+///
+/// Holds a [`NerBackend`] and a [`LanguageDetector`] (both required)
+/// plus an optional [`Tokenizer`]. The default [`analyze`] entrypoint
+/// matches Microsoft Presidio's `AnalyzerEngine` ordering: detect
+/// language, run NER (with the detected language as a hint),
+/// tokenize, derive keywords.
+///
+/// When the caller already knows the language — e.g. a document
+/// uploaded with explicit metadata — use [`analyze_in_language`] to
+/// bypass detection.
+///
+/// Construct via [`builder`].
+///
+/// [`NerBackend`]: crate::ner::NerBackend
+/// [`LanguageDetector`]: crate::language::LanguageDetector
+/// [`Tokenizer`]: crate::tokenizer::Tokenizer
+/// [`analyze`]: Self::analyze
+/// [`analyze_in_language`]: Self::analyze_in_language
+/// [`builder`]: Self::builder
+pub struct Engine {
+    pub(super) ner: Arc<dyn NerBackend>,
+    pub(super) language: Arc<dyn LanguageDetector>,
+    pub(super) tokenizer: Option<Arc<dyn Tokenizer>>,
+}
+
+impl Engine {
+    /// Start building an engine.
+    pub fn builder() -> EngineBuilder {
+        EngineBuilder::default()
+    }
+
+    /// Run all configured components, detecting the language from
+    /// `text` first.
+    pub async fn analyze(&self, text: &str) -> Result<Artifacts> {
+        let detection = self.language.detect(text)?;
+        self.run(text, detection).await
+    }
+
+    /// Run all configured components with the caller-asserted
+    /// `language`, bypassing detection.
+    ///
+    /// Use this when the language is known a priori (uploaded with
+    /// metadata, set by a UI selector, etc.). The asserted language
+    /// is attached to [`Artifacts::language`] and carries
+    /// [`LanguageProvenance::Asserted`] internally so downstream
+    /// code can distinguish it from a detector-produced result.
+    pub async fn analyze_in_language(
+        &self,
+        text: &str,
+        language: LanguageTag,
+    ) -> Result<Artifacts> {
+        let detection = Some(LanguageDetection {
+            language,
+            confidence: None,
+            provenance: LanguageProvenance::Asserted,
+        });
+        self.run(text, detection).await
+    }
+
+    async fn run(&self, text: &str, detection: Option<LanguageDetection>) -> Result<Artifacts> {
+        let language_hint = detection.as_ref().map(|d| &d.language);
+        let entities = self.ner.recognize(text, language_hint).await?;
+        let tokens = match &self.tokenizer {    
+            Some(t) => Some(t.tokenize(text)?),
+            None => None,
+        };
+        let keywords = tokens.as_deref().map(derive_keywords);
+        let language = detection.map(|d| d.language);
+
+        Ok(Artifacts {
+            entities,
+            language,
+            tokens,
+            keywords,
+        })
+    }
+}
+
+impl std::fmt::Debug for Engine {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        f.debug_struct("Engine")
+            .field("tokenizer", &self.tokenizer.is_some())
+            .finish_non_exhaustive()
+    }
+}
+
+fn derive_keywords(tokens: &[Token]) -> HashSet<String> {
+    tokens
+        .iter()
+        .filter(|t| !t.is_stop)
+        .map(|t| t.text.to_lowercase())
+        .collect()
+}
 
 #[cfg(test)]
 mod tests {
@@ -21,10 +125,10 @@ mod tests {
     use crate::ner::NoopNerBackend;
     use crate::tokenizer::UnicodeTokenizer;
 
-    fn english_engine() -> NlpEngine {
+    fn english_engine() -> Engine {
         let tags = ["en".parse().unwrap()];
         let det = LinguaLanguageDetector::for_languages(&tags).unwrap();
-        NlpEngine::builder()
+        Engine::builder()
             .with_ner(NoopNerBackend)
             .with_language_detector(det)
             .build()
@@ -61,7 +165,7 @@ mod tests {
         let det = LinguaLanguageDetector::for_languages(&tags).unwrap();
         let lang: LanguageTag = "en".parse().unwrap();
         let tok = UnicodeTokenizer::with_language(&lang).unwrap();
-        let engine = NlpEngine::builder()
+        let engine = Engine::builder()
             .with_ner(NoopNerBackend)
             .with_language_detector(det)
             .with_tokenizer(tok)
